(front): rename bookdetails to bookDetails
