+++ frontend/app/(protected)/(drawer)/(tabs)/bookDetails.tsx
